removing uclust demo
since this code is not making use of qiime directly, it is not an acceptable use of the qiime-licensed version of uclust (in my interpretation, haven't contacted R.E. about it). instead, we'll present the usearch version of the code, which makes sense anyway since the uclust tax assigner that i hacked together here is now implemented and tested in QIIME. this notebook is intended to show how to test one that is not implemented in QIIME (just because we want it to be clear that QIIME is not a requirement for this).